fix: fixing warnings to keep the standardization of code
