+++ lib/pages/GamePage/CustomWidgets/UsersDropdown.dart
@@ -1,16 +1,17 @@
-import 'package:bizinuca/models/UserModel.dart';
+import 'package:bizinuca/models/user_model.dart';
 import 'package:flutter/material.dart';
 
 class UsersDropDown extends StatelessWidget {
+  const UsersDropDown(this.user, this.usersDropdown, this.onChange);
+
   final UserModel user;
   final List<DropdownMenuItem<UserModel>> usersDropdown;
   final Function onChange;
 
-  UsersDropDown(this.user, this.usersDropdown, this.onChange);
   @override
   Widget build(BuildContext context) {
     return DropdownButton<UserModel>(
-        style: TextStyle(
+        style: const TextStyle(
           color: Colors.black,
           fontWeight: FontWeight.bold,
         ),
