+++ lib/pages/GamePage/GamePage.dart
@@ -1,5 +1,5 @@
 import 'package:bizinuca/components/primary_button.dart';
-import 'package:bizinuca/models/UserModel.dart';
+import 'package:bizinuca/models/user_model.dart';
 import 'package:bizinuca/services/FeedBackService.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
